requirements.txt
dicttoxml==1.6.6
falcon==0.3.0
gunicorn==19.4.5
Jinja2==2.8
lupa==1.2
MarkupSafe==0.23
-e git+https://github.com/shitara/mongoengine.git@15ec26a93a2209610473fb2da975afc56fee7827#egg=mongoengine
pathtools==0.1.2
pycrypto==2.6.1
pymongo==3.2.1
python-mimeparse==1.5.1
PyYAML==3.11
six==1.10.0
watchdog==0.8.3
